Fix int vs NSInteger warnings for 64 bit iOS and other deprecation warnings.

+++ G-licious/Controllers/StationsViewController.m
@@ -189,7 +189,7 @@
 }
 
 #pragma mark - TTSlidingPageController delegate methods
-- (int)numberOfPagesForSlidingPagesViewController:(TTScrollSlidingPagesController *)source
+- (NSInteger)numberOfPagesForSlidingPagesViewController:(TTScrollSlidingPagesController *)source
 {
     return self.menu.count;
 }
